requirements.txt
# ==========================================
# Semantic Router Dependencies
# ==========================================

# Data Generation & Environment Configuration
openai>=1.30.0
python-dotenv==1.0.1

# NLP, Embeddings, and Array Processing
sentence-transformers==2.5.1
numpy==1.26.4

# Machine Learning & Model Serialization
scikit-learn==1.4.1.post1
joblib==1.3.2
xgboost==2.0.3

# API Infrastructure & Validation
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.3

# Visualization & Interactive Demo
matplotlib==3.8.3
jupyter==1.0.0